% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{CI_calculator}
\alias{CI_calculator}
\title{Confidence Interval Calculator}
\usage{
CI_calculator(fit)
}
\arguments{
\item{fit}{A linreg_fit object.}
}
\value{
A data frame containing coefficient estimates, lower, and upper 95\% CI.
}
\description{
Calculates 95\% confidence intervals for model coefficients.
}
\examples{
model <- linreg_fit(mpg ~ wt + hp + disp, data = mtcars)
CI_calculator(model)
}
