% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{data_info}
\alias{data_info}
\title{Compute variable statistics}
\usage{
data_info(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A list with mean, median, standard deviation, min, and max.
}
\description{
Calculates basic summary statistics for a numeric variable.
}
\examples{
data_info(mtcars$mpg)
}
