% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg_predict}
\alias{linreg_predict}
\title{Predict from fitted linear regression model}
\usage{
linreg_predict(fit, newdata)
}
\arguments{
\item{fit}{A linreg_fit object.}

\item{newdata}{A data frame containing predictor values for prediction.
Must include all predictor variables used in the fitted model.}
}
\value{
A numeric vector of predicted values.
}
\description{
Uses the fitted model from linreg_fit() to predict Y values given new data.
This function can be used, for example, to predict a new patient's blood
pressure (BLP) using the fitted regression model.
}
\examples{
model <- linreg_fit(mpg ~ wt + hp + disp, data = mtcars)
new_data <- data.frame(wt = 2.5, hp = 110, disp = 109)
linreg_predict(model, new_data)
}
